src/requirements.txt
pandas
langchain_openai
langchain_core
langgraph
langchain
python-dotenv
argparse
numpy==1.23.5
scikit-learn
fastapi
uvicorn
python-multipart
temporalio
httpx